Fix negative duration in Activity
Fix overflow due to cast from int to short while extending the maximum
duration to the one in PatternActivity.

+++ mobitopp/src/main/java/edu/kit/ifv/mobitopp/simulation/activityschedule/linkedlist/ActivitySequenceAsLinkedList.java
@@ -94,8 +94,8 @@ public class ActivitySequenceAsLinkedList
 																		(byte)nrWithinDay,
 																		activityType,
 																		startDate,
-																		(short) duration,
-																		(short) tripDuration
+																		duration,
+																		tripDuration
 																);
 			return activity;
 		}
